--- frontend/pages/SoftwarePage/SoftwareAddPage/SoftwareCustomPackage/AddSoftwareCustomPackageForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddSoftwareCustomPackageForm";
