+++ frontend/pages/SoftwarePage/SoftwareAddPage/SoftwareCustomPackage/SoftwareCustomPackage.tsx
@@ -1,10 +1,13 @@
-import React, { useContext } from "react";
+import React, { useContext, useEffect } from "react";
 import { InjectedRouter } from "react-router";
-import { Location } from "history";
 
 import PATHS from "router/paths";
 import { buildQueryStringFromParams, QueryParams } from "utilities/url";
-import softwareAPI from "services/entities/software";
+import softwareAPI, {
+  MAX_FILE_SIZE_BYTES,
+  MAX_FILE_SIZE_MB,
+  UPLOAD_TIMEOUT,
+} from "services/entities/software";
 import { LEARN_MORE_ABOUT_BASE_LINK } from "utilities/constants";
 import { NotificationContext } from "context/notification";
 import { getErrorReason } from "interfaces/errors";
@@ -14,22 +17,13 @@ import CustomLink from "components/CustomLink";
 import PackageForm from "pages/SoftwarePage/components/PackageForm";
 import { IPackageFormData } from "pages/SoftwarePage/components/PackageForm/PackageForm";
 
-import { ISoftwareAddPageQueryParams } from "../SoftwareAddPage";
-import AddSoftwareCustomPackageForm from "./AddSoftwareCustomPackageForm";
-import { ICustomPackageAppFormData } from "./AddSoftwareCustomPackageForm/AddSoftwareCustomPackageForm";
 import { getErrorMessage } from "./helpers";
 
 const baseClass = "software-custom-package";
 
-// 8 minutes + 15 seconds to account for extra roundtrip time.
-export const UPLOAD_TIMEOUT = (8 * 60 + 15) * 1000;
-export const MAX_FILE_SIZE_MB = 500;
-export const MAX_FILE_SIZE_BYTES = MAX_FILE_SIZE_MB * 1024 * 1024;
-
 interface ISoftwarePackageProps {
   currentTeamId: number;
   router: InjectedRouter;
-  location: Location<ISoftwareAddPageQueryParams>;
   isSidePanelOpen: boolean;
   setSidePanelOpen: (isOpen: boolean) => void;
 }
@@ -37,12 +31,38 @@ interface ISoftwarePackageProps {
 const SoftwareCustomPackage = ({
   currentTeamId,
   router,
-  location,
   isSidePanelOpen,
   setSidePanelOpen,
 }: ISoftwarePackageProps) => {
   const { renderFlash } = useContext(NotificationContext);
   const [isUploading, setIsUploading] = React.useState(false);
+
+  useEffect(() => {
+    let timeout: NodeJS.Timeout;
+
+    const beforeUnloadHandler = (e: BeforeUnloadEvent) => {
+      e.preventDefault();
+      // Next line with e.returnValue is included for legacy support
+      // e.g.Chrome / Edge < 119
+      e.returnValue = true;
+    };
+
+    // set up event listener to prevent user from leaving page while uploading
+    if (isUploading) {
+      addEventListener("beforeunload", beforeUnloadHandler);
+      timeout = setTimeout(() => {
+        removeEventListener("beforeunload", beforeUnloadHandler);
+      }, UPLOAD_TIMEOUT);
+    } else {
+      removeEventListener("beforeunload", beforeUnloadHandler);
+    }
+
+    // clean up event listener and timeout on component unmount
+    return () => {
+      removeEventListener("beforeunload", beforeUnloadHandler);
+      clearTimeout(timeout);
+    };
+  }, [isUploading]);
 
   const onCancel = () => {
     router.push(
@@ -126,12 +146,6 @@ const SoftwareCustomPackage = ({
         onCancel={onCancel}
         onSubmit={onSubmit}
       />
-      {/* <AddSoftwareCustomPackageForm
-        showSchemaButton={!isSidePanelOpen}
-        onClickShowSchema={() => setSidePanelOpen(true)}
-        onCancel={onCancel}
-        onSubmit={onSubmit}
-      /> */}
     </div>
   );
 };
