--- frontend/pages/SoftwarePage/components/AddPackage/_styles.scss
@@ -1,3 +1,0 @@
-.add-package {
-  margin-top: $pad-large;
-}
