--- frontend/pages/SoftwarePage/components/AddPackage/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddPackage";
